refactor(simulation): move delete flow into feature module
